requirements.txt
Django==4.2.3
django_silk==5.0.3
psycopg==3.1.9
Faker==19.2.0
ipython==8.14.0
djangorestframework==3.14.0
django_filter==23.2
black==23.7.0
isort==5.12.0
